src/requirements.txt
mysql-connector-python
Pillow
pillow-heif